+++ components/descriptions/item.ts
@@ -4,24 +4,26 @@ import template from './item.vdt';
 import { useConfigContext } from '../config';
 
 export interface DescriptionItemProps {
-    label?: string
+    label?: string | number | VNode
     content?: string | number | VNode
-    option?: string | number | VNode
+    append?: string | number | VNode
 }
 
-export interface DescriptionBlocks {
-    option: null
+export interface DescriptionItemBlocks {
+    append: null
+    label: null
 }
 
+export interface DescriptionItemEvents {}
 
 const typeDefs: Required<TypeDefs<DescriptionItemProps>> = {
-    label: String,
+    label: [String, Number, VNode],
     content: [String, Number, VNode],
-    option: [String, Number, VNode]
+    append: [String, Number, VNode]
 };
 
 
-export class DescriptionItem extends Component<DescriptionItemProps, DescriptionBlocks> {
+export class DescriptionItem extends Component<DescriptionItemProps, DescriptionItemEvents, DescriptionItemBlocks> {
     static template = template;
     static typeDefs = typeDefs;