+++ packages/kpc-vue-next/__tests__/components/editable.spec.ts
@@ -35,7 +35,7 @@ describe('Editable', () => {
         container.querySelector<HTMLDivElement>('.k-icon')!.click();
         await wait();
 
-        const input = container.querySelector('.k-input input') as HTMLInputElement;
+        const input = container.querySelector('.k-input textarea') as HTMLTextAreaElement;
         expect(input.selectionStart).to.eql(0);
         expect(input.selectionEnd).to.eql(1);